requirements.txt
jinja2>=3.0.0
markdown>=3.4.0
pyyaml>=6.0.0
pillow>=9.0.0
anthropic>=0.18.0
